[TSql] Move files & classes according to new layout
