+++ pmd-tsql/src/main/antlr4/net/sourceforge/pmd/lang/tsql/ast/TSqlLexer.g4
