+++ pmd-tsql/src/main/java/net/sourceforge/pmd/lang/tsql/ast/package-info.java
@@ -2,4 +2,4 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.tsql.antlr4;
+package net.sourceforge.pmd.lang.tsql.ast;
