+++ pmd-tsql/src/main/java/net/sourceforge/pmd/lang/tsql/cpd/TSqlLanguage.java
@@ -2,7 +2,9 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.cpd;
+package net.sourceforge.pmd.lang.tsql.cpd;
+
+import net.sourceforge.pmd.cpd.AbstractLanguage;
 
 /**
  * @author pguyot@kallisys.net
