+++ pmd-tsql/src/main/resources/META-INF/services/net.sourceforge.pmd.cpd.Language
@@ -1,1 +1,1 @@
-net.sourceforge.pmd.cpd.TSqlLanguage
+net.sourceforge.pmd.lang.tsql.cpd.TSqlLanguage
