--- pmd-tsql/src/main/resources/META-INF/services/net.sourceforge.pmd.lang.Language
@@ -1,1 +1,0 @@
-net.sourceforge.pmd.lang.tsql.TSqlLanguageModule
